+++ src/lib/TermNode.cc
@@ -1,13 +1,13 @@
 #include "TermNode.h"
 
-bool TermNode::AddLetters(const std::string &word, uint8_t letter) {
+bool TermNode::AddLetters(const std::string &word, uint16_t letter) {
 	return true;
 }
 
-bool TermNode::TryAddLetters(const std::string &word, uint8_t letter) {
+bool TermNode::TryAddLetters(const std::string &word, uint16_t letter) {
 	return true;
 }
 
-bool TermNode::TryLetters(const std::string &word, uint8_t letter) {
+bool TermNode::TryLetters(const std::string &word, uint16_t letter) {
 	return true;
 }
