+++ src/lib/TrieNode.h
@@ -5,6 +5,7 @@
 #include <cstdint>
 #include <algorithm>
 #include <memory>
+#include <iostream>
 #include <unordered_map>
 #include "Node.h"
 #include "TermNode.h"
@@ -17,9 +18,9 @@ class TrieNode : public Node {
 		bool AddWord(const std::string &word);
 		bool TryWord(const std::string &word);
 
-		virtual bool AddLetters(const std::string &word, uint8_t letter);
-		virtual bool TryAddLetters(const std::string &word, uint8_t letter);
-		virtual bool TryLetters(const std::string &word, uint8_t letter);
+		virtual bool AddLetters(const std::string &word, uint16_t letter);
+		virtual bool TryAddLetters(const std::string &word, uint16_t letter);
+		virtual bool TryLetters(const std::string &word, uint16_t letter);
 
 	private:
 		std::unordered_map<char,std::unique_ptr<Node>> child_;
